requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
pyodbc
pandas
openpyxl
requests
itsdangerous
